Got rid of a couple of warnings in the example

+++ example/example.cpp
@@ -6,7 +6,7 @@
 
 int main() {
 	rlcg::ReversibleLCG rng(42);
-	const unsigned int numtests = 10000;
+	const int numtests = 10000;
 
 	std::vector<unsigned int> forward;
 	std::cout << "\nForward:\n";
@@ -19,7 +19,7 @@ int main() {
 	for(int i = numtests - 2; i>=0; --i){
 		unsigned int val = rng.prev();
 		std::cout << val << std::endl;
-		assert(val == forward[i]);
+		assert(val == forward[static_cast<unsigned int>(i)]);
 	}
 
 	std::vector<unsigned int> backward;
@@ -32,7 +32,7 @@ int main() {
 	for(int i = numtests - 2; i>=0; --i){
 		unsigned int val = rng.next();
 		std::cout << val << std::endl;
-		assert(val == backward[i]);
+		assert(val == backward[static_cast<unsigned int>(i)]);
 	}
 	return 0;
 }
